+++ packages/cli/src/utils/registry/index.ts
@@ -1,8 +1,8 @@
 import path from "node:path";
 import process from "node:process";
 import * as v from "valibot";
-import fetch from "node-fetch";
-import { HttpsProxyAgent } from "https-proxy-agent";
+import { fetch } from "node-fetch-native";
+import { createProxy } from "node-fetch-native/proxy";
 import { error } from "../errors.js";
 import { getEnvProxy } from "../get-env-proxy.js";
 import type { Config } from "../get-config.js";
@@ -10,29 +10,19 @@ import * as schemas from "./schema.js";
 
 const baseUrl = process.env.COMPONENTS_REGISTRY_URL ?? "https://shadcn-svelte.com";
 
-export type RegistryItem = v.Output<typeof schemas.registryItemSchema>;
+export type RegistryItem = v.InferOutput<typeof schemas.registryItemSchema>;
 
 export async function getRegistryIndex() {
 	try {
 		const [result] = await fetchRegistry(["index.json"]);
 
 		return v.parse(schemas.registryIndexSchema, result);
-	} catch (e) {
+	} catch {
 		throw error(`Failed to fetch components from registry.`);
 	}
 }
 
-export async function getRegistryStyles() {
-	try {
-		const [result] = await fetchRegistry(["styles/index.json"]);
-
-		return v.parse(schemas.stylesSchema, result);
-	} catch (e) {
-		throw error(`Failed to fetch styles from registry.`);
-	}
-}
-
-export async function getRegistryBaseColors() {
+export function getBaseColors() {
 	return [
 		{
 			name: "slate",
@@ -57,18 +47,31 @@ export async function getRegistryBaseColors() {
 	];
 }
 
+export function getStyles() {
+	return [
+		{
+			name: "default",
+			label: "Default",
+		},
+		{
+			name: "new-york",
+			label: "New York",
+		},
+	];
+}
+
 export async function getRegistryBaseColor(baseColor: string) {
 	try {
 		const [result] = await fetchRegistry([`colors/${baseColor}.json`]);
 
 		return v.parse(schemas.registryBaseColorSchema, result);
-	} catch (e) {
+	} catch {
 		throw error(`Failed to fetch base color from registry.`);
 	}
 }
 
-type RegistryIndex = v.Output<typeof schemas.registryIndexSchema>;
-export async function resolveTree(index: RegistryIndex, names: string[]) {
+type RegistryIndex = v.InferOutput<typeof schemas.registryIndexSchema>;
+export async function resolveTree(index: RegistryIndex, names: string[], includeRegDeps = true) {
 	const tree: RegistryIndex = [];
 
 	for (const name of names) {
@@ -80,7 +83,7 @@ export async function resolveTree(index: RegistryIndex, names: string[]) {
 
 		tree.push(entry);
 
-		if (entry.registryDependencies) {
+		if (includeRegDeps && entry.registryDependencies) {
 			const dependencies = await resolveTree(index, entry.registryDependencies);
 			tree.push(...dependencies);
 		}
@@ -98,14 +101,14 @@ export async function fetchTree(config: Config, tree: RegistryIndex) {
 		const result = await fetchRegistry(paths);
 
 		return v.parse(schemas.registryWithContentSchema, result);
-	} catch (e) {
+	} catch {
 		throw error(`Failed to fetch tree from registry.`);
 	}
 }
 
 export function getItemTargetPath(
 	config: Config,
-	item: Pick<v.Output<typeof schemas.registryItemWithContentSchema>, "type">,
+	item: Pick<v.InferOutput<typeof schemas.registryItemWithContentSchema>, "type">,
 	override?: string
 ) {
 	// Allow overrides for all items but ui.
@@ -125,17 +128,19 @@ export function getItemTargetPath(
 
 async function fetchRegistry(paths: string[]) {
 	const proxyUrl = getEnvProxy();
-	const agent = proxyUrl ? new HttpsProxyAgent(proxyUrl) : undefined;
+	const proxy = proxyUrl ? createProxy({ url: proxyUrl }) : {};
 	try {
 		const results = await Promise.all(
 			paths.map(async (path) => {
-				const response = await fetch(`${baseUrl}/registry/${path}`, { agent });
+				const response = await fetch(`${baseUrl}/registry/${path}`, {
+					...proxy,
+				});
 				return await response.json();
 			})
 		);
 
 		return results;
-	} catch (e) {
+	} catch {
 		throw error(`Failed to fetch registry from ${baseUrl}.`);
 	}
 }