+++ packages/cli/src/utils/sveltekit.ts
@@ -1,13 +1,20 @@
+import fs from "node:fs";
+import path from "node:path";
 import { execa } from "execa";
-import { getPackageManager } from "./get-package-manager.js";
+import { detect } from "package-manager-detector";
+import { COMMANDS } from "package-manager-detector/agents";
 import { loadProjectPackageInfo } from "./get-package-info.js";
 
-// if it's a SvelteKit project, run sync so that the aliases are always up to date
+// if it's a SvelteKit project, run `svelte-kit sync` if the `.svelte-kit` dir is missing
 export async function syncSvelteKit(cwd: string) {
 	const isSvelteKit = isUsingSvelteKit(cwd);
 	if (isSvelteKit) {
-		const packageManager = await getPackageManager(cwd);
-		await execa(packageManager === "npm" ? "npx" : packageManager, ["svelte-kit", "sync"], {
+		// we'll exit early since syncing is rather slow
+		if (fs.existsSync(path.join(cwd, ".svelte-kit"))) return;
+
+		const { agent } = await detect({ cwd });
+		const [pm] = COMMANDS[agent ?? "npm"].agent.split(" ") as [string];
+		await execa(pm === "npm" ? "npx" : pm, ["svelte-kit", "sync"], {
 			cwd,
 		});
 	}